use redux and do this thing right
