+++ frontend/src/actions/constants.js
@@ -1,0 +1,2 @@
+export const OPEN_CHAT = 'OPEN_CHAT'
+export const CLOSE_CHAT = 'CLOSE_CHAT'
