+++ frontend/src/reducers/index.js
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux'
+
+import map from './map'
+//import notifications from './notifications'
+
+export default combineReducers({
+  map//,
+  //notifications
+})
