+++ frontend/src/reducers/map/filters.js
@@ -1,0 +1,21 @@
+const initialState = {
+  dataForPresentation: {
+    metacodes: {},
+    mappers: {},
+    synapses: {}
+  },
+  filters: {
+    metacodes: [],
+    mappers: [],
+    synapses: []
+  },
+  visible: {
+    metacodes: [],
+    mappers: [],
+    synapses: []
+  }
+}
+
+export default function (state = initialState, action) {
+  return state
+}
