+++ frontend/src/reducers/map/importDialogue.js
@@ -1,0 +1,5 @@
+const initialState = {}
+
+export default function (state = initialState, action) {
+  return state
+}
