+++ frontend/src/reducers/map/selected.js
@@ -1,0 +1,8 @@
+const initialState = {
+  Nodes: [],
+  Edges: []
+}
+
+export default function (state = initialState, action) {
+  return state
+}
