+++ frontend/src/reducers/map/synapseCard.js
@@ -1,0 +1,3 @@
+export default function (state = null, action) {
+  return state
+}
