+++ frontend/src/reducers/map/topicCard.js
@@ -1,0 +1,3 @@
+export default function (state = null, action) {
+  return state
+}
