+++ frontend/src/reducers/notifications/index.js
@@ -1,0 +1,3 @@
+export default function notifications(state = [], action) {
+  return state
+}
