+++ package.json
@@ -49,8 +49,9 @@
     "react-draggable": "3.0.3",
     "react-dropzone": "4.1.2",
     "react-onclickoutside": "6.5.0",
+    "react-redux": "^5.0.6",
     "react-router": "3.0.5",
-    "redux": "3.7.2",
+    "redux": "^3.7.2",
     "riek": "1.1.0",
     "simplewebrtc": "2.2.2",
     "socket.io": "1.3.7",
